requirements.txt
groq
transformers
sentence-transformers
langchain
langchain-groq
langchain-huggingface
requests
pandas
numpy
python-dotenv
langchain-community
psycopg2-binary
pgvector
Flask
PyYAML
gunicorn
flask_cors